Move Azure backend test into another file.
